encapsulation constness some compiler warnings

+++ src/layers/lbann_target_layer_distributed_minibatch_parallel_io.cpp
@@ -68,7 +68,7 @@ void lbann::target_layer_distributed_minibatch_parallel_io::setup(int num_prev_n
   }
 
   /// @todo put in warning about bad target size
-  if(num_prev_neurons != m_num_neurons) {
+  if(static_cast<uint>(num_prev_neurons) != m_num_neurons) {
     stringstream err;
     err << __FILE__ << " " << __LINE__
         << " ::  lbann_target_layer_distributed_minibatch_parallel_io: number of neurons in previous layer (" << num_prev_neurons << ") does not match the number of neurons in the target layer (" << m_num_neurons <<  ")";
